requirements.txt
django==4.0.*
djangorestframework==3.13.*
drf_yasg==1.20.*